requirements.txt
PyMuPDF>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0
genanki>=0.13.0
click>=8.1.0
httpx>=0.28.0
rich>=13.0.0
